Put Windows Service, Connection Settings in Same View.

+++ asagiv.datapush.common/Interfaces/IClientConnectionSettings.cs
@@ -2,7 +2,7 @@
 {
     public interface IClientConnectionSettings
     {
-        public long Id { get; }
+        public uint Id { get; }
         public string ConnectionName { get; }
         public string ConnectionString { get; }
         public bool IsPullNode { get; }
